# api/openapi.yaml
openapi: 3.0.3
info:
  title: GEX Faker API
  version: 2.0.0
  description: |
    Historical GEX data server with per-API-key sequential playback.
    Serves JSONL data files with configurable cache modes.
servers:
  - url: /
    description: Current server
...
paths:
...
  /{ticker}/classic/{aggregation}:
    get:
      operationId: getClassicGexChain
      summary: Get GEX chain data
      description: |
        Provides a ladder-style histogram of GEX by Open Interest and Volume.
        Call GEX is netted against Put GEX for each strike.
        Includes historical "lookback" data points and major calculated levels.
      tags: [GexBot parity]
      parameters:
        - name: ticker
          in: path
          required: true
          description: Ticker symbol (e.g., SPX)
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
          example: SPX
        - name: aggregation
          in: path
          required: true
          description: "Aggregation period: full, zero, or one"
          schema:
            type: string
            enum: [gex_full, gex_zero, gex_one]
          example: gex_zero
      responses:
        '200':
          description: GEX chain data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GexData'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data not found or exhausted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /{ticker}/state/{type}:
    get:
      operationId: getStateProfile
      summary: Get state profile data (GEX or Greeks)
      description: |
        Unified endpoint for state profile data.

        When type is an aggregation (full, zero, one):
          Returns GEX profile data - nets out imbalanced calls against imbalanced puts
          to locate areas of significant imbalanced exposure.

        When type is a greek (delta_zero, gamma_zero, etc.):
          Returns Greek profile data - shows net imbalance of transactions for the
          specified Greek, allowing analysis of customer long vs. short positioning.
      tags: [GexBot parity]
      parameters:
        - name: ticker
          in: path
          required: true
          description: Ticker symbol (e.g., SPX)
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
          example: SPX
        - name: type
          in: path
          required: true
          description: "Aggregation period (full, zero, one) OR Greek type (delta_zero, gamma_zero, etc.)"
          schema:
            type: string
            enum: [gex_full, gex_zero, gex_one, delta_zero, gamma_zero, delta_one, gamma_one, charm_zero, vanna_zero, charm_one, vanna_one]
          example: gex_zero
      responses:
        '200':
          description: Profile data (GexData for aggregations, GreekProfileData for greeks)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/GexData'
                  - $ref: '#/components/schemas/GreekProfileData'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data not found or exhausted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{ticker}/orderflow/orderflow:
    get:
      operationId: getOrderflowLatest
      summary: Get latest orderflow metrics
      description: |
        Provides a real-time snapshot of key orderflow metrics for the nearest
        and next expiries. Includes aggregated and net dealer exposure (DEX),
        GEX, convexity, vanna, and charm, as well as orderflow indicators.
      tags: [GexBot parity]
      parameters:
        - name: ticker
          in: path
          required: true
          description: Ticker symbol (e.g., SPX)
          schema:
            type: string
            pattern: '^[A-Z_]{1,10}$'
          example: SPX
      responses:
        '200':
          description: Orderflow metrics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderflowData'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Data not found or exhausted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /options/{ticker}/expiries:
    get:
      operationId: getOptionsExpiries
      summary: List valid option expiries
      description: |
        Returns valid option expiration dates for a ticker over the published
        horizon: the loaded data date through +90 calendar days. Near-term
        expiries are grounded in the loaded data's min_dte/sec_min_dte; the rest
        follow the options calendar (daily near-term for index/major-ETF tickers,
        then weekly). Remove dashes to form explicit-expiry WebSocket group
        suffixes (2026-07-17 -> 20260717).
      tags: [GexBot parity]
      parameters:
        - name: ticker
          in: path
          required: true
          description: Canonical option-owning ticker (e.g., SPX, SPY, QQQ)
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
          example: SPX
      responses:
        '200':
          description: Valid option expiries within the horizon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpiriesResponse'
        '400':
          description: Unsupported or invalid option ticker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No expiries available (current date is not a calendar date)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /load:
    post:
      operationId: load
      summary: Load a day or a span of trading days for replay
      description: |
        **Faker-only.** Materializes (if needed) and loads one day or a contiguous span as a single
        continuous dataset, so seek/replay can cross day boundaries. Asynchronous: returns a job id
        immediately; poll GET /load/status/{jobId} until state=done. Provide exactly one of: `date`
        (a single day), `from`+`to` (the archived days within the inclusive span), or an explicit
        `dates` list. Gated by STUDIO_AUTH_TOKEN when the server has one set.
      tags: [Faker control plane]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadRequest'
      responses:
        '202':
          description: Load job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadStatus'
        '400':
          description: Invalid request or no archived dates in range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /available/{date}:
    get:
      operationId: getAvailable
      summary: What data exists for a date
      description: |
        **Faker-only.** Hierarchical listing of the data available for a date —
        tickers, packages, and categories present in the data directory.
      tags: [Faker control plane]
      parameters:
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          example: "2025-12-04"
        - name: ticker
          in: query
          required: false
          description: Filter to a specific ticker
          schema:
            type: string
            pattern: '^[A-Z]{1,5}$'
          example: SPX
      responses:
        '200':
          description: Available data for the date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableResponse'
...
  /reset:
    post:
      operationId: reset
      summary: Rewind playback positions
      description: "**Faker-only.** Rewinds all or one API key's replay cursor to the first snapshot. Gated by STUDIO_AUTH_TOKEN when the server has one set."
      tags: [Faker control plane]
      parameters:
        - name: key
          in: query
          required: false
          description: Reset only this API key (omit for all)
          schema:
            type: string
      responses:
        '200':
          description: Positions reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetResponse'
  /seek:
    post:
      operationId: seek
      summary: Seek playback positions to a timestamp
      description: |
        **Faker-only.** Sets a key's cursor to the first snapshot at or after the given unix timestamp.
        In range mode it resolves across the whole loaded span (resolved_ts, day, in_gap, clamped).
        Per-client — never token-gated.
      tags: [Faker control plane]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeekRequest'
      responses:
        '200':
          description: Seek successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeekResponse'
        '400':
          description: Invalid request or timestamp out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    GexData:
      type: object
      required:
        - timestamp
        - ticker
      properties:
        timestamp:
          type: integer
          format: int64
          example: 1764340202
        ticker:
          type: string
          example: SPX
        min_dte:
          type: integer
          example: 0
        sec_min_dte:
          type: integer
          example: 3
        spot:
          type: number
          format: double
          example: 6822.95
        zero_gamma:
          type: number
          format: double
          example: 0
        major_pos_vol:
          type: number
          format: double
          example: 0
        major_pos_oi:
          type: number
          format: double
          example: 0
        major_neg_vol:
          type: number
          format: double
          example: 0
        major_neg_oi:
          type: number
          format: double
          example: 0
        strikes:
          type: array
          items: {}
        sum_gex_vol:
          type: number
          format: double
          example: 0
        sum_gex_oi:
          type: number
          format: double
          example: 43565.23
        delta_risk_reversal:
          type: number
          format: double
          example: 0
        max_priors:
          type: array
          items: {}
...
    ExpiriesResponse:
      type: object
      required: [ticker, start_date, end_date, expiries]
      properties:
        ticker:
          type: string
          description: Canonical ticker used for the expiry lookup
          example: SPX
        start_date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Start of the published horizon (the loaded data date)
          example: 2026-08-07
        end_date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: End of the horizon (start date + 90 calendar days)
          example: 2026-11-05
        expiries:
          type: array
          items:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Valid option expiries in YYYY-MM-DD within the horizon
...
    ResetResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: All cache positions reset to index 0
        count:
          type: integer
          description: Number of positions reset
          example: 12
...
    LoadRequest:
      type: object
      description: >
        Provide exactly one of: date (a single day); from+to (inclusive span — the available archived
        days within it are loaded); or an explicit dates list.
      properties:
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Load a single day.
          example: '2026-08-07'
        from:
          type: string
          example: '2026-08-06'
        to:
          type: string
          example: '2026-08-10'
        dates:
          type: array
          items:
            type: string
          example: ['2026-08-06', '2026-08-07']
...
    LoadStatus:
      type: object
      properties:
        job_id:
          type: string
          example: range-1
        dates:
          type: array
          items:
            type: string
        state:
          type: string
          enum: [queued, running, done, error]
        done:
          type: integer
          description: Days materialized so far
        total:
          type: integer
          description: Days in the span
        loaded_range:
          $ref: '#/components/schemas/LoadedRange'
        error:
          type: string
...
    GreekProfileData:
      type: object
      required:
        - timestamp
        - ticker
      properties:
        timestamp:
          type: integer
          format: int64
          example: 1764340202
        ticker:
          type: string
          minLength: 2
          maxLength: 5
          example: SPX
        spot:
          type: number
          format: double
          example: 6822.95
        min_dte:
          type: integer
          example: 0
        sec_min_dte:
          type: integer
          example: 3
        major_positive:
          type: number
          format: double
          example: 6100
        major_negative:
          type: number
          format: double
          example: 5800
        major_long_gamma:
          type: number
          format: double
          example: 6585
        major_short_gamma:
          type: number
          format: double
          example: 6585
        mini_contracts:
          type: array
          items:
            type: array
          description: Array of contract data arrays
    OrderflowData:
      type: object
      required:
        - timestamp
        - ticker
      properties:
        timestamp:
          type: integer
          format: int64
        ticker:
          type: string
          minLength: 2
          maxLength: 5
        spot:
          type: number
          format: double
        z_mlgamma:
          type: number
        z_msgamma:
          type: number
        o_mlgamma:
          type: number
        o_msgamma:
          type: number
        zero_mcall:
          type: number
        zero_mput:
          type: number
        one_mcall:
          type: number
        one_mput:
          type: number
        zcvr:
          type: number
        ocvr:
          type: number
        zgr:
          type: number
        ogr:
          type: number
        zvanna:
          type: number
        ovanna:
          type: number
        zcharm:
          type: number
        ocharm:
          type: number
        agg_dex:
          type: number
        one_agg_dex:
          type: number
        agg_call_dex:
          type: number
        one_agg_call_dex:
          type: number
        agg_put_dex:
          type: number
        one_agg_put_dex:
          type: number
        net_dex:
          type: number
        one_net_dex:
          type: number
        net_call_dex:
          type: number
        one_net_call_dex:
          type: number
        net_put_dex:
          type: number
        one_net_put_dex:
          type: number
        dexoflow:
          type: number
        gexoflow:
          type: number
        cvroflow:
          type: number
        one_dexoflow:
          type: number
        one_gexoflow:
          type: number
        one_cvroflow:
          type: number
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid ticker symbol
...
    AvailableResponse:
      type: object
      properties:
        date:
          type: string
          description: The requested date
          example: "2025-12-04"
        tickers:
          type: array
          items:
            $ref: '#/components/schemas/TickerData'
          description: Available tickers with their packages
        summary:
          $ref: '#/components/schemas/DataSummary'
...